+++ src/pages/User/PetFormImagePreview.jsx
@@ -1,4 +1,5 @@
 import React from 'react';
+import { useFormikContext } from 'formik';
 import Typography from '@mui/material/Typography';
 import Box from '@mui/material/Box';
 import IconButton from '@mui/material/IconButton';
@@ -9,6 +10,8 @@ const PetFormImageReview = function PetFormImageReviewComponent({
   image,
   handleDelete,
 }) {
+  const { isSubmitting } = useFormikContext();
+
   // If the old (uploaded) images are about to be deleted, then don't show it on
   // the preview section.
   if (image.action && image.action === 'delete') return undefined;
@@ -29,6 +32,7 @@ const PetFormImageReview = function PetFormImageReviewComponent({
           color="error"
           sx={{ position: 'absolute', top: 0, right: 0 }}
           onClick={() => handleDelete(image)}
+          disabled={isSubmitting}
         >
           <FontAwesomeIcon icon="fa-solid fa-circle-xmark" />
         </IconButton>
